+++ blobmanager.js
@@ -3,15 +3,15 @@ class BlobManager {
 		this.gameboard = gameboard;
 
 		this.blobs = [];
-		this.initialBlobAmount = 20;
+		this.initialBlobAmount = 30;
 		this.standardBlobSize = 20;
 
 		colorMode(HSB, 255, 255, 255);
-		this.playerBlobColor = color(1, 255, 170);
+		this.playerBlobColor = Constants.PLAYER_BLOB_COLOR;
 
 		this.initBlobs();
 
-		this.deadBlobs = new FIFOQueue();	
+		this.deadBlobs = new FIFOQueue();
 	}
 
 	update() {
@@ -32,23 +32,42 @@ class BlobManager {
 	}
 
 	addBlob() {
-		let pos = createVector(random(this.gameboard.width), random(this.gameboard.height));
-		let vel = createVector(randomGaussian(0, 2), randomGaussian(0, 2));
-		let size = randomGaussian(this.standardBlobSize, 0.1);
-		let col = color(random(255), 120, 230);
+		let pos = createVector(
+			random(
+				this.gameboard.width), 
+			random(
+				this.gameboard.height));
+		let vel = createVector(
+			randomGaussian(0, 2), 
+			randomGaussian(0, 2));
+		let size = randomGaussian(
+			this.standardBlobSize, 0.1);
+		let col = color(
+			random(255), 
+			Constants.BLOB_SATURATION, 
+			Constants.BLOB_BRIGHTNESS);
 		let inputModule = new PerlinInput();
 		let isManual = false;
-		let blob = new Blob(size, col, pos, vel, isManual);
+		let blob = new Blob(
+			size, 
+			col, 
+			pos, 
+			vel, 
+			isManual);
 		blob.setInputModule(inputModule);
 		this.blobs.push(blob);
 	}
 
-	initPlayerBlob() {
-		let pos = createVector(width/2, height/2);
+	initPlayerBlob(pos) {
 		let vel = createVector(0, 0);
 		let size = this.standardBlobSize;
 		let isManual = true;
-		this.playerBlob = new Blob(size, this.playerBlobColor, pos, vel, isManual);
+		this.playerBlob = new Blob(
+			size, 
+			this.playerBlobColor, 
+			pos, 
+			vel, 
+			isManual);
 		this.blobs.push(this.playerBlob);
 		return this.playerBlob;
 	}
@@ -106,6 +125,9 @@ class BlobManager {
 			let deadBlob = this.deadBlobs.poll();
 			let index = this.blobs.indexOf(deadBlob);
 			this.blobs.splice(index, 1);
+
+			// Add new for every dead one removed
+			this.addBlob();
 		}
 	}