+++ constants.js
@@ -1,0 +1,13 @@
+class Constants {
+	static get BLOB_SATURATION() {
+		return 120;
+	}
+
+	static get BLOB_BRIGHTNESS() {
+		return 230;
+	}
+	
+	static get PLAYER_BLOB_COLOR() {
+		return color(1, 255, 170);
+	}	
+}