+++ followcam.js
@@ -1,24 +1,32 @@
 class FollowCam {
-	constructor(blobManager) {
-		this.blobManager = blobManager;
-		this.lookingAtGameSpacePos = null;
+	constructor(gameBoard, allBlobs) {
+		this.gameBoard = gameBoard;
+
+		this.allBlobs = allBlobs;
 		this.followee = null;
 		this.viewZeroInGameSpace = null;
 		this.gameZeroInScreenSpace = null;
+
+		this.lookingAtGameSpacePos = null;
+		this.movementDamping = 0.01;
+		this.lastFrameTime = millis();
 	}
 
 	follow(entity) {
 		this.followee = entity;
+		this.lookingAtGameSpacePos = entity.pos;
 	}
 
 	update() {
 		if (this.followee) {
-			this.lookingAtGameSpacePos = this.followee.pos;
+			this.smoothFollow(this.followee);
 			this.viewZeroInGameSpace = this.getViewZeroInGameSpace(this.lookingAtGameSpacePos);
 			this.gameZeroInScreenSpace = this.convertToScreenSpace(this.viewZeroInGameSpace);
 		}
 
 		this.renderObjectsInView();
+
+		this.lastFrameTime = millis();
 	}
 
 	getViewZeroInGameSpace(lookingAtPos) {
@@ -28,14 +36,20 @@ class FollowCam {
 	}
 
 	renderObjectsInView() {
+		this.render(this.gameBoard);
+
 		let thisHandle = this;
-		let everythingThisCamRendereds = this.getRenderedObjects();
-		everythingThisCamRendereds.forEach(function(rendObj) {
-			if (thisHandle.isWithinViewFrustum(rendObj)) {
-				let screenSpaceCoord = thisHandle.convertToScreenSpace(rendObj.pos);
-				rendObj.displayAt(screenSpaceCoord);
-			}
+		let allBlobs = this.allBlobs;
+		allBlobs.forEach(function(rendObj) {
+			// if (thisHandle.isWithinViewFrustum(rendObj)) {
+				thisHandle.render(rendObj);
+			// }
 		});
+	}
+
+	render(object) {
+		let screenSpaceCoord = this.convertToScreenSpace(object.pos);
+		object.displayAt(screenSpaceCoord);
 	}
 
 	convertToScreenSpace(gameSpaceCoord) {
@@ -46,15 +60,30 @@ class FollowCam {
 		return p5.Vector.add(screenSpaceCoord, this.viewZeroInGameSpace);
 	}
 
-	getRenderedObjects() {
-		return this.blobManager.allBlobs;
+	// isWithinViewFrustum(object) {
+	// 	let screenSpaceCoord = this.convertToScreenSpace(object.pos);
+	// 	return screenSpaceCoord.x >= 0 && 
+	// 	screenSpaceCoord.x < width &&
+	// 	screenSpaceCoord.y >= 0 && 
+	// 	screenSpaceCoord.y < height;
+	// }
+
+	smoothFollow(entity) {
+		if (!entity.pos) {
+			console.log("FollowCam: Entity has no position to smoothfollow");
+			return;
+		}
+		let desiredPos = entity.pos;
+		let currentPos = this.lookingAtGameSpacePos;	
+		let nextPos = p5.Vector.lerp(
+			currentPos, desiredPos, 
+			this.getSecondsSinceLastFrame() * 
+			this.movementDamping * 
+			desiredPos.dist(currentPos));
+		this.lookingAtGameSpacePos = nextPos;
 	}
 
-	isWithinViewFrustum(object) {
-		let screenSpaceCoord = this.convertToScreenSpace(object.pos);
-		return screenSpaceCoord.x >= 0 && 
-		screenSpaceCoord.x < width &&
-		screenSpaceCoord.y >= 0 && 
-		screenSpaceCoord.y < height;
-	} 
+	getSecondsSinceLastFrame() {
+		return (millis() - this.lastFrameTime) / 1000;
+	}
 }